Alm Execution: adding reports in build summary Alm Lab Management Execution: adding reports in build summary FS Execution: removing LR and Mobile content

+++ UFTBuildTask/AlmLabManagementTask/localTask.ps1
@@ -15,7 +15,8 @@ param(
 	[string] $varUseCDA,
 	[string] $varDeploymentAction,
 	[string] $varDeploymentEnvironmentName,
-	[string] $varDeprovisioningAction
+	[string] $varDeprovisioningAction,
+	[string][Parameter(Mandatory=$false)] $varReportName
 )
 
 $uftworkdir = $env:UFT_LAUNCHER
@@ -24,9 +25,19 @@ $stdout = "$uftworkdir\temp_build.log"
 $stderr = "$uftworkdir\temp_error_build.log"
 $jar = """$uftworkdir\bin\hpe.application.automation.tfs.almrestrunner-1.0-jar-with-dependencies.jar"""
 
-$args = "-jar $jar AlmLabManagement  ""serv:$varAlmserv"" ""user:$varUserName"" ""domain:$varDomain"" ""project:$varProject"" ""testSet:$varTestSet"" ""timeSlotDuration:$varTimeslotDuration"" ""pass:$varPass"" ""runType:$varRunType"" ""envconfID:$varEnvironmentConfigurationID"" ""desc:$varDescription"" ""useCDA:$varUseCDA"" ""deploymentAction:$varDeploymentAction"" ""depEnvName:$varDeploymentEnvironmentName"" ""deprovisioningAction:$varDeprovisioningAction"""
+if (-Not $varReportName)
+{
+	$varReportName = "ALM Lab Management Report"
+}
+$args = "-jar $jar AlmLabManagement  ""serv:$varAlmserv"" ""user:$varUserName"" ""domain:$varDomain"" ""project:$varProject"" ""testSet:$varTestSet"" ""timeSlotDuration:$varTimeslotDuration"" ""pass:$varPass"" ""runType:$varRunType"" ""envconfID:$varEnvironmentConfigurationID"" ""desc:$varDescription"" ""useCDA:$varUseCDA"" ""deploymentAction:$varDeploymentAction"" ""depEnvName:$varDeploymentEnvironmentName"" ""deprovisioningAction:$varDeprovisioningAction"" ""repname:$varReportName"""
 
 echo $args
+
+$report = Join-Path $env:UFT_LAUNCHER -ChildPath "res\$($varReportName)"
+if (Test-Path $report)
+{
+	Remove-Item $report
+}
 
 $process = (Start-Process java -ArgumentList $args -RedirectStandardOutput $stdout -RedirectStandardError $stderr -PassThru -Wait)
 
@@ -43,5 +54,16 @@ else
 	Get-Content $stderr
 }
 
-Remove-Item $stdout
-Remove-Item $stderr
+if (Test-Path $report)
+{
+	Write-Host "##vso[task.uploadsummary]"$report
+}
+
+if (Test-Path $report)
+{
+	Remove-Item $stdout
+}
+if (Test-Path $report)
+{
+	Remove-Item $stderr
+}
